update_kernel.sh: fix unified version file updates
The previous commit broke the kernel-version.mk automated update, since the
kernel version files are now split. However, older branches still use the
unified file, so compatibility must be kept.

Check for the presence of the kernel version-specific file. If it doesn't exist,
assume we're using the unified file and do the version update accordingly.

Fixes: cbb9d08 ("update_kernel.sh: update it to new kernel hash/version file way")

Tested-by: Petr Štetiar <ynezz@true.cz>
Signed-off-by: Rui Salvaterra <rsalvaterra@gmail.com>